filter: extract rotation domain from cache_sensitive.rs
#6237 asked to split filter/engine/cache_sensitive.rs (732 LOC) by
call-frequency domain. Hostile plan review (Codex + research) converged
on NARROWING to a rotation-only extraction rather than the issue's
literal 3-way split: the cached TX-selection rebuild path and the input
counter-descriptor capture are two products of the same seed/replay
descriptor workflow, so splitting them apart would fragment a cohesive
unit. This change therefore extracts ONLY the cold filter-rotation
semantic-comparison domain and keeps the two hot-path domains together
in the parent module.

Pure behavior-preserving code motion. The rotation domain
(three_color_policer_semantics_match, filter_term_semantics_match with
the #5293 exhaustive FilterTerm destructure, dscp_sensitive_filter_
semantics_match, the input_dscp / input_per_packet_l4 family-changed
predicates, and the interface_input/output has_dscp / has_per_packet_l4
lookups) plus its colocated cache_sensitive_2400_tests move VERBATIM
into a new engine/cache_sensitive/rotation.rs submodule (Rust 2018
file + sibling-directory layout).

The parent cache_sensitive.rs now declares `mod rotation;` and
re-exports the six pub(crate) rotation functions, so engine/mod.rs's
eight-symbol re-export block and every downstream consumer stay
untouched. Function bodies are byte-identical with the pre-split file;
the only change in the moved region is that the module-level
`use super::super::*` (and the test module's glob) gain one `super::`
because the submodule sits one level deeper. The shared engine helpers
(matching::term_matches_v4/v6, eval::filter_log_match) are pub(in engine)
and remain visible to the descendant module, so no visibility widening
is needed.

The #5293 compile-time anti-drift guard — the exhaustive
`let FilterTerm { .. } = new;` destructure with no `..` rest pattern —
moves intact and stays exhaustive, so adding a FilterTerm field still
fails to compile until it is classified. That compile-time guard, not
file co-location, is what makes physical separation from the rebuild
path safe (superseding the original #1546 co-location rationale).

Validation: cargo build --release and cargo clippy --release are clean
(the sole clippy error is a pre-existing mut_from_ref deny in untouched
afxdp/umem/mmap.rs:150). Full cargo test --release --test-threads=1 is
green (4134 passed / 2 ignored / 0 failed, plus sibling test binaries);
the three #5293 comparator tests now run under
filter::engine::cache_sensitive::rotation::cache_sensitive_2400_tests.
A diff of the moved src range 276-732 against origin/master confirms
byte-identical bodies (only the +1 `super::` on the two glob imports
differs). filter/README.md rotation-symbol references repointed to
rotation.rs; the cached-TX reference stays on cache_sensitive.rs.

Closes #6237
